requirements.txt
aiogram>=3.0
google-api-python-client
google-auth
python-dotenv
